chore: rename editable-bio to bio-editor
Refs: #285
