+++ src/components/pages/account-page/current-user-info/index.tsx
@@ -13,7 +13,7 @@ import { HttpError } from '@/utils/error/custom/http-error'
 import { generateRedirectLoginPath } from '@/utils/login-path/generate-redirect-login-path.server'
 import { ChangePasswordLink } from './change-password-link'
 import { EditableAvatar, LoadingEditableAvatar } from './editable-avatar'
-import { EditableBio } from './editable-bio'
+import { EditableBio } from './editable-bio/bio-editor'
 import { EditableEmail } from './editable-email'
 import { EditableName, LoadingEditableName } from './editable-name'
 import { PrivateModeSwitch } from './private-mode-switch'
